## 펜 그리기 요소 송수신 및 제어 구현 완료 * EditorVC - DisplayView - DecorateObjectManager 관계 재정의 * DisplayView, DecorateObjectManager Delegate Protocol 재정의 * DisplayView의 하위 뷰 관리방식 재정의 및 메소드 명칭 변경 * DecorateObjectManager의 데이터 관리 방식 재정의 및 메소드 명칭 변경 * Category 이용하여 UIView 확장한 뒤, stringTag 사용할 수 있도록 함. 이를 이용하여 각 뷰에 identifier 부여

+++ wos_intern_project/wos_intern_project/PECropRectView.m
@@ -230,7 +230,7 @@
         
         if (self.keepingAspectRatio) {
             CGRect constrainedRect;
-            if (fabsf(resizeControlView.translation.x) < fabsf(resizeControlView.translation.y)) {
+            if (fabsf((float)resizeControlView.translation.x) < fabsf((float)resizeControlView.translation.y)) {
                 constrainedRect = [self constrainedRectWithRectBasisOfHeight:rect aspectRatio:self.fixedAspectRatio];
             } else {
                 constrainedRect = [self constrainedRectWithRectBasisOfWidth:rect aspectRatio:self.fixedAspectRatio];
@@ -246,7 +246,7 @@
                           CGRectGetHeight(self.initialRect) - resizeControlView.translation.y);
         
         if (self.keepingAspectRatio) {
-            if (fabsf(resizeControlView.translation.x) < fabsf(resizeControlView.translation.y)) {
+            if (fabsf((float)resizeControlView.translation.x) < fabsf((float)resizeControlView.translation.y)) {
                 rect = [self constrainedRectWithRectBasisOfHeight:rect aspectRatio:self.fixedAspectRatio];
             } else {
                 rect = [self constrainedRectWithRectBasisOfWidth:rect aspectRatio:self.fixedAspectRatio];
@@ -260,7 +260,7 @@
         
         if (self.keepingAspectRatio) {
             CGRect constrainedRect;
-            if (fabsf(resizeControlView.translation.x) < fabsf(resizeControlView.translation.y)) {
+            if (fabsf((float)resizeControlView.translation.x) < fabsf((float)resizeControlView.translation.y)) {
                 constrainedRect = [self constrainedRectWithRectBasisOfHeight:rect aspectRatio:self.fixedAspectRatio];
             } else {
                 constrainedRect = [self constrainedRectWithRectBasisOfWidth:rect aspectRatio:self.fixedAspectRatio];
@@ -275,7 +275,7 @@
                           CGRectGetHeight(self.initialRect) + resizeControlView.translation.y);
         
         if (self.keepingAspectRatio) {
-            if (fabsf(resizeControlView.translation.x) < fabsf(resizeControlView.translation.y)) {
+            if (fabsf((float)resizeControlView.translation.x) < fabsf((float)resizeControlView.translation.y)) {
                 rect = [self constrainedRectWithRectBasisOfHeight:rect aspectRatio:self.fixedAspectRatio];
             } else {
                 rect = [self constrainedRectWithRectBasisOfWidth:rect aspectRatio:self.fixedAspectRatio];
